Add optional tracing for DNS lookup, TCP connect, TLS negotiation
Use -Ddd.trace.connection.details.enabled=true to enable tracing
for connections details.

This is an PoC. It has no tests, a number of things that need cleaned
up (XXX, FIXME, TODO), and likely needs tested and adjusted to work
broadly with various JVM versions (in particular the TLS negotiation
instrumentation, but also look for variation in the DNS lookup and
TCP connect private mtehods).